+++ config.json
@@ -1,21 +1,23 @@
 {
-  "realm": "cim",
+  "realm": "university",
   "auth-server-url": "http://192.168.1.204:8080/auth/",
   "ssl-required": "external",
-  "resource": "unified-admin",
+  "resource": "agent-manager",
   "verify-token-audience": false,
   "credentials": {
-    "secret": "27080cdf-cdd8-4db1-b3ee-fdb0669b0222"
+    "secret": "QOJS4gyCMzv09GPs5ik0AvllBM05VIAC"
   },
   "use-resource-role-mappings": true,
   "confidential-port": 0,
   "policy-enforcer": {},
-  "CLIENT_ID": "unified-admin",
-  "CLIENT_DB_ID": "95536d4e-c5d5-4876-8cc3-99025e18fc60",
+  "CLIENT_ID": "agent-manager",
+  "CLIENT_DB_ID": "QOJS4gyCMzv09GPs5ik0AvllBM05VIAC",
   "GRANT_TYPE": "password",
   "GRANT_TYPE_PAT": "client_credentials",
-  "USERNAME_ADMIN": "uadmin",
-  "PASSWORD_ADMIN": "uadmin",
-  "SCOPE_NAME": "Any deafult scope",
-  "bearer-only":true
+  "USERNAME_ADMIN": "admin",
+  "PASSWORD_ADMIN": "admin",
+  "SCOPE_NAME": "Any default scope",
+  "bearer-only": true,
+  "finesse-server-url": "https://uccx-morocco.ucce.ipcc:8445",
+  "GROUP": "Agent"
 }
