+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ef-keycloak-connect",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Node JS keycloak adapter for authentication and authorization.",
   "main": "index.js",
   "scripts": {
@@ -15,6 +15,7 @@
   "dependencies": {
     "axios": "^0.21.0",
     "express": "^4.17.1",
+    "express-session": "^1.17.2",
     "keycloak-connect": "^10.0.2"
   }
 }
